+++ barretenberg/cpp/src/barretenberg/vm2/simulation/gadgets/sha256.cpp
@@ -86,7 +86,7 @@ void Sha256::compression(MemoryInterface& memory,
                          MemoryAddress output_addr)
 {
     uint32_t execution_clk = execution_id_manager.get_execution_id();
-    uint32_t space_id = memory.get_space_id();
+    uint16_t space_id = memory.get_space_id();
 
     // Default values are FF(0) as that is what the circuit would expect
     std::array<MemoryValue, 8> state;
